refactor: RequestHandler에는 HttpRequest 관련 로직만 있도록 handleHttpRequest 이동
