+++ src/main/java/webserver/RequestHandler.java
@@ -1,26 +1,25 @@
 package webserver;
 
-import annotation.RequestMappingHandler;
 import http.HttpRequest;
-import http.HttpResponse;
+import http.Mime;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import util.StringUtils;
 
 import java.io.BufferedReader;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.net.Socket;
 import java.util.Map;
-import java.util.Objects;
 
 public class RequestHandler implements Runnable {
     private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
 
     private final Socket connection;
+    private final ResponseHandler responseHandler;
 
     public RequestHandler(Socket connectionSocket) {
         this.connection = connectionSocket;
+        this.responseHandler = new ResponseHandler(connection);
     }
 
     @Override
@@ -34,10 +33,7 @@ public class RequestHandler implements Runnable {
             // 요청 읽기
             HttpRequest httpRequest = new HttpRequest(reader);
             printLogs(httpRequest);
-            HttpResponse httpResponse = handleHttpRequest(httpRequest);
-
-            ResponseHandler responseHandler = new ResponseHandler(connection);
-            responseHandler.response(httpResponse);
+            responseHandler.response(httpRequest);
         } catch (Throwable e) {
             logger.error(e.getMessage());
         }
@@ -51,24 +47,5 @@ public class RequestHandler implements Runnable {
         logger.debug("Method : {}, URI : {}, Version : {}", httpRequest.method(), httpRequest.uri(), httpRequest.version());
         logger.debug("Headers : {}", requestBuilder);
         logger.debug("Mime : {}, Body : {}", httpRequest.mime(), httpRequest.getBody());
-    }
-
-    private HttpResponse handleHttpRequest(HttpRequest httpRequest) {
-        String path = httpRequest.uri().getPath();
-        String extension = StringUtils.getExtension(path);
-        HttpResponse httpResponse;
-        if (!Objects.equals(extension, path)) {
-            // 정적 파일 응답
-            httpResponse = HttpResponse.ok(path, httpRequest.mime());
-        } else {
-            // 잘못된 http request이면 /error.html response 생성
-            try {
-                httpResponse = RequestMappingHandler.invokeMethod(httpRequest);
-            } catch (Throwable e) {
-                logger.error("메소드를 실행하는데 오류가 발생했습니다.\n{}", (Object) e.getStackTrace());
-                httpResponse = HttpResponse.notFound();
-            }
-        }
-        return httpResponse;
     }
 }
